requirements.txt
azure-servicebus==0.50.1
azure-cosmos==3.1.2
azure-functions==1.0.4
azure-keyvault==1.0.0
azure-storage-blob==12.1.0
msrestazure==0.6.2
xmltodict==0.12.0
azure-mgmt-resource==2.2.0
requests==2.22.0
pandas